requirements.txt
aiogram==2.25.1
python-dotenv
requests
aiohttp
asyncio
urllib3
pytz==2024.1